+++ Riot/Modules/Application/LegacyAppDelegate.m
@@ -233,6 +233,17 @@ NSString *const AppDelegateUniversalLinkDidChangeNotification = @"AppDelegateUni
 @property (nonatomic, assign, getter=isRoomListDataReady) BOOL roomListDataReady;
 
 /**
+ An observer token for `RecentsViewControllerDataReadyNotification`s notifications.
+ */
+@property (nonatomic, nullable) id roomListDataReadyObserver;
+
+/**
+ An optional completion block that will be called when a `RecentsViewControllerDataReadyNotification`
+ is observed during app launch.
+ */
+@property (nonatomic, copy, nullable) void (^roomListDataReadyCompletion)(void);
+
+/**
  Flag to indicate whether a cache clear is being performed.
  */
 @property (nonatomic, assign, getter=isClearingCache) BOOL clearingCache;
@@ -2374,6 +2385,7 @@ NSString *const AppDelegateUniversalLinkDidChangeNotification = @"AppDelegateUni
                 case MXSessionStateInitialised:
                 case MXSessionStateBackgroundSyncInProgress:
                     self.roomListDataReady = NO;
+                    [self listenForRoomListDataReady];
                     isLaunching = YES;
                     break;
                 case MXSessionStateStoreDataReady:
@@ -2411,7 +2423,7 @@ NSString *const AppDelegateUniversalLinkDidChangeNotification = @"AppDelegateUni
             return;
         }
         
-        [self ensureRoomListDataReadyWithCompletion:^{
+        void (^finishAppLaunch)(void) = ^{
             [self hideLaunchAnimation];
             
             if (self.setPinCoordinatorBridgePresenter)
@@ -2443,7 +2455,17 @@ NSString *const AppDelegateUniversalLinkDidChangeNotification = @"AppDelegateUni
                 // Enable listening of incoming key verification requests
                 [self enableIncomingKeyVerificationObserver:mainSession];
             }
-        }];
+        };
+        
+        if (self.isRoomListDataReady)
+        {
+            finishAppLaunch();
+        }
+        else
+        {
+            // An observer has been set in didFinishLaunching that will call the stored block when ready
+            self.roomListDataReadyCompletion = finishAppLaunch;
+        }
     }
 }
 
@@ -2599,29 +2621,31 @@ NSString *const AppDelegateUniversalLinkDidChangeNotification = @"AppDelegateUni
     [self handleAppState];
 }
 
-/**
- Ensures room list data is ready.
- 
- @param completion Completion block to be called when it's ready. Not dispatched in case the data is already ready.
- */
-- (void)ensureRoomListDataReadyWithCompletion:(void(^)(void))completion
+- (void)listenForRoomListDataReady
 {
-    if (self.isRoomListDataReady)
+    if (self.roomListDataReadyObserver)
     {
-        completion();
+        return;
     }
-    else
-    {
-        NSNotificationCenter * __weak notificationCenter = [NSNotificationCenter defaultCenter];
-        __block id observer = [[NSNotificationCenter defaultCenter] addObserverForName:RecentsViewControllerDataReadyNotification
-                                                                                object:nil
-                                                                                 queue:[NSOperationQueue mainQueue]
-                                                                            usingBlock:^(NSNotification * _Nonnull notification) {
-            [notificationCenter removeObserver:observer];
-            self.roomListDataReady = YES;
-            completion();
-        }];
-    }
+    
+    MXWeakify(self);
+    NSNotificationCenter * __weak notificationCenter = [NSNotificationCenter defaultCenter];
+    self.roomListDataReadyObserver = [[NSNotificationCenter defaultCenter] addObserverForName:RecentsViewControllerDataReadyNotification
+                                                                            object:nil
+                                                                             queue:[NSOperationQueue mainQueue]
+                                                                        usingBlock:^(NSNotification * _Nonnull notification) {
+        MXStrongifyAndReturnIfNil(self);
+        
+        [notificationCenter removeObserver:self.roomListDataReadyObserver];
+        self.roomListDataReady = YES;
+        self.roomListDataReadyObserver = nil;
+        
+        if (self.roomListDataReadyCompletion)
+        {
+            self.roomListDataReadyCompletion();
+            self.roomListDataReadyCompletion = nil;
+        }
+    }];
 }
 
 #pragma mark -