+++ Riot/Modules/MatrixKit/Views/Authentication/MXKAuthenticationRecaptchaWebView.m
@@ -102,7 +102,7 @@ var onloadCallback = function() { \
     if ([urlString hasPrefix:@"js:"])
     {
         // Listen only to scheme of the JS-WKWebView bridge
-        NSString *jsonString = [[[urlString componentsSeparatedByString:@"js:"] lastObject]  stringByAddingPercentEncodingWithAllowedCharacters:[NSCharacterSet URLQueryAllowedCharacterSet]];
+        NSString *jsonString = [[[urlString componentsSeparatedByString:@"js:"] lastObject] stringByRemovingPercentEncoding];
         NSData *jsonData = [jsonString dataUsingEncoding:NSUTF8StringEncoding];
 
         NSError *error;