+++ Riot/Modules/Room/Search/DataSources/RoomSearchDataSource.m
@@ -96,6 +96,10 @@
             {
                 continueBlock();
             }
+            else if (result.result.unsignedData.relations.thread)
+            {
+                continueBlock();
+            }
             else
             {
                 [roomDataSource.room liveTimeline:^(id<MXEventTimeline> liveTimeline) {
@@ -143,8 +147,9 @@
             {
                 if (cellData.hasThreadRoot)
                 {
-                    MXThread *thread = cellData.bubbleComponents.firstObject.thread;
-                    ThreadSummaryView *threadSummaryView = [[ThreadSummaryView alloc] initWithThread:thread];
+                    id<MXThreadProtocol> thread = cellData.bubbleComponents.firstObject.thread;
+                    ThreadSummaryView *threadSummaryView = [[ThreadSummaryView alloc] initWithThread:thread
+                                                                                             session:self.mxSession];
                     [bubbleCell.tmpSubviews addObject:threadSummaryView];
 
                     threadSummaryView.translatesAutoresizingMaskIntoConstraints = NO;