+++ RiotSwiftUI/Modules/Room/LocationSharing/View/LocationSharingView.swift
@@ -66,6 +66,9 @@ struct LocationSharingView: View {
                     }
                 }
                 .navigationBarTitleDisplayMode(.inline)
+                .introspectNavigationController { navigationController in
+                    ThemeService.shared().theme.applyStyle(onNavigationBar: navigationController.navigationBar)
+                }
                 .ignoresSafeArea()
                 .alert(item: $context.alertInfo) { info in
                     if let secondaryButton = info.secondaryButton {
@@ -89,9 +92,6 @@ struct LocationSharingView: View {
         .accentColor(theme.colors.accent)
         .activityIndicator(show: context.viewState.showLoadingIndicator)
         .navigationViewStyle(StackNavigationViewStyle())
-        .introspectNavigationController { navigationController in
-            ThemeService.shared().theme.applyStyle(onNavigationBar: navigationController.navigationBar)
-        }
     }
     
     @ViewBuilder